pyproject.toml
[tool.poetry]
name = "flashcards-aiobot"
version = "0.1.0"
description = ""
authors = ["0xcitadel@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
aiogram = "^3.3.0"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
greenlet = "^3.0.3"
aiogram-dialog = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
